Add Material-UI
I've wrapped the <App> component in a themeProvider HOC from the Material-UI library, this can be configured to use any of the provided color variables internal to the library, designed to match Material Design specifications. Alternatively, it should be possible to pass hex values directly into the theme object. In addition to this, each component can be styled and further customized according to that component's API.